+++ lib/belated/worker.rb
@@ -18,25 +18,9 @@ class Belated
 
         break if job == :shutdown
 
-        log call_job(job)
-      end
-    end
-
-    # rubocop:disable Lint/RescueException
-    def call_job(job)
-      if job.respond_to?(:call)
-        job.call
-      else
+        log "Worker #{@number} got job: #{job.inspect}"
         job.perform
       end
-    rescue Exception => e
-      case e.class
-      when Interrupt, SignalException
-        raise e
-      else
-        e.inspect
-      end
     end
-    # rubocop:enable Lint/RescueException
   end
 end