+++ spec/belated/queue_spec.rb
@@ -1,0 +1,5 @@
+RSpec.describe Belated::Queue do
+  it 'logs the completed jobs' do
+    # queue = Belated::Queue.new
+  end
+end