+++ spec/druby_job_spec.rb
@@ -11,7 +11,7 @@ RSpec.describe Belated do
       Belated.config.workers = 1
       thread = Thread.new { Belated.new }
       dummy = DummyWorker.new
-      dummy.queue.push(DumDum.new)
+      dummy.queue.push(Belated::JobWrapper.new(job: DumDum.new))
       sleep 0.05
       expect(dummy.queue.empty?).to be_truthy
       thread.kill